lxd/storage: Remove legacy lvm implementation
